feat: improve KASLR (#99)

+++ kernel/src/kconfig.rs
@@ -17,7 +17,7 @@ const fn parse_log_lvl(s: &str) -> log::Level {
 
 /// The size of the stack in pages
 #[kconfig_declare::symbol("kernel.stack-size-pages")]
-pub const STACK_SIZE_PAGES: usize = 128;
+pub const STACK_SIZE_PAGES: u32 = 128;
 
 /// The size of the trap handler stack in pages
 #[kconfig_declare::symbol("kernel.trap-stack-size-pages")]
@@ -25,7 +25,7 @@ pub const TRAP_STACK_SIZE_PAGES: usize = 16;
 
 /// The size of the kernel heap in pages
 #[kconfig_declare::symbol("kernel.heap-size-pages")]
-pub const HEAP_SIZE_PAGES: usize = 8192; // 32 MiB
+pub const HEAP_SIZE_PAGES: u32 = 8192; // 32 MiB
 
 // TODO: This should be configurable
 #[allow(non_camel_case_types)]
